Major commit: Drop MULTIDPY support.
Compiz used to have theoretical support for multiple X servers.

This is now removed to simplify the code.

This change has zero effects on the user.

So, now there is only one CompDisplay, accessed using the global
variable "display", declared in fusilli-core.h and defined in
display.c.

The plugin API(initDisplay etc.) is not changed in this commit.

Notes:
display.c: removeDisplay.
	Remove screens *before* removing display!
	Add hack in forEachDisplayObject, avoid crashing during pushing of core plugin

core.c:
	Remove display *after* unloading plugins.